fix: added missing files
