+++ src/renderer/marketplace/notarization/index.jsx
@@ -4,8 +4,8 @@ import { connect } from 'react-redux';
 import { marketplaceOperations } from 'common/marketplace';
 import { NotarizationDetailsContainer } from './details/notarization-details-container';
 import { RequestNotarization } from './process/request-notarization';
-import { NotarizationTOC } from './common/toc';
-import { NotarizationTOCdisagreement } from './common/toc-disagreement';
+import { NotarizationTOC } from './common/toc-container';
+import { NotarizationTOCdisagreement } from './common/toc-disagreement-container';
 
 class MarketplaceNotariesComponent extends Component {
 	async componentDidMount() {
